+++ src/components/ProfileImageUploader.jsx
@@ -1,14 +1,19 @@
-import { useState, forwardRef } from "react";
+import { useState, forwardRef, useImperativeHandle } from "react";
 import styled from "styled-components";
 
 const ProfileImageUploader = forwardRef((props, ref) => {
-  const {
-    prev,
-    setIsCancelled,
-    setSelectedFile,
-    previewImage,
-    setPreviewImage,
-  } = props;
+  const { prev } = props;
+  const [isCancelled, setIsCancelled] = useState(false);
+  const [selectedFile, setSelectedFile] = useState(null);
+  const [previewImage, setPreviewImage] = useState(null);
+
+  useImperativeHandle(ref, () => ({
+    getProfileImageData: () => ({
+      isCancelled,
+      selectedFile,
+      previewImage,
+    }),
+  }));
 
   const handleImageChange = (e) => {
     const file = e.target.files[0];
@@ -22,13 +27,19 @@ const ProfileImageUploader = forwardRef((props, ref) => {
       };
       reader.readAsDataURL(file);
       setSelectedFile(file);
-      // } else {
-      //   setPreviewImage(null);
+    } else {
+      setPreviewImage(null);
     }
     e.target.value = null;
   };
 
   const handleThumbnailImage = () => ref.current.click();
+
+  const handleProfileImageCancel = () => {
+    setIsCancelled(true);
+    setSelectedFile(null);
+    setPreviewImage(prev);
+  };
 
   return (
     <>
@@ -43,11 +54,15 @@ const ProfileImageUploader = forwardRef((props, ref) => {
         alt="Profile thumbnail"
         onClick={handleThumbnailImage}
       />
+      <button type="button" onClick={handleProfileImageCancel}>
+        프사취소
+      </button>
     </>
   );
 });
 
 export default ProfileImageUploader;
+
 
 const StyledProfileImageUploader = {
   Input: styled.input`
