+++ src/helpers/regexp.js
@@ -1,5 +1,6 @@
 export const REGEXP = {
-  namd: /^[a-zA-Z0-9ㄱ-ㅎㅏ-ㅣ가-힣]{1,15}$/, // 영문, 숫자, 한글 1~15자
+  // namd: /^[a-zA-Z0-9ㄱ-ㅎㅏ-ㅣ가-힣]{1,15}$/, // 영문, 숫자, 한글 1~15자
+  name: /^[a-zA-Zㄱ-ㅎㅏ-ㅣ가-힣]+([ ][a-zA-Zㄱ-ㅎㅏ-ㅣ가-힣]+){0,14}$/,  // 한글, 영문
 
   phone:
     /^(?:(?:01(?:0|1|[6-9])-(?:\d{3,4})-\d{4})|(?:02|0[3-6][1-4]|01(?:0|1|[6-9]))-(?:\d{3,4})-\d{4})$/,
