--- src/pages/ChildManage/ele/ChildProfile.jsx
@@ -1,5 +1,0 @@
-const ChildProfile = () => {
-  return <>dd</>;
-};
-
-export default ChildProfile;
